Move TryPyrefly.module.css to inline css
Summary: I'm starting to work on making changes to add support for running python code and I noticed that we aren't using inline css with stylex in the `TryPyrefly` code. This diff makes this migration so that future css will be in the same file and be easier to read.

Reviewed By: stroxler, ndmitchell

Differential Revision: D73127286

fbshipit-source-id: 6aaeb963f5cda13562f2fa947beb0e1fc593fcc1