Sort modules by name when collecting errors
Summary:
I noticed that type errors are not in sorted order today — they are sorted within a module, but when there are more than 1 module then the order in which errors are reported for those modules are not stable.

I’m thinking we should have a stable ordering for modules. Not only does this align with what Pyre1 does, but also if we run pyrefly multiple times on the same project and errors are reported in the same order, it makes it easier to locate specific errors in subsequent runs.

Reviewed By: ndmitchell

Differential Revision: D73162725

fbshipit-source-id: 4f0dceb4fbddc99a69c84b5568cbb2ed5e3e7232