Eliminate `shift_remove` from attribute adding
Summary:
A challenge I ran into building out the `TypeInfo` datatype is that `NarrowedAttr`
has an enum, which made it nontrivial to add a mutation-based `add_narrow`, which
in turn nudged me toward a `shift_remove` based implemntation of `add_narrow` in
`NarrowedAttrs`, since I needed to destroy an owned copy of a `NarrowedAttr` in
the recursive case.

With `mem::swap` I can avoid this - by making a dummy `NarrowedAttr` value and
using it to take ownership of `self`, we can make a mutation for `add_narrow`.

This also gets rid of a reordering of the attributes that was midly annoying,
because it made it harder to guess at the order in which narrows occurred
when looking at a dumped type_info, now the ordering is more intuitive in my
opinion :)

Reviewed By: ndmitchell

Differential Revision: D73148333

fbshipit-source-id: 3e8eac64265c5e5921c84417defbbdc3f40c3603