Remove TODO, I expeimented and won't make the change
Summary:
I tried changing `NarrowedAttr` to a tuple of options, and I broke things
catastrophically by messing up an invariant.

The fact that my first attempt to do anything with a simpler and less-safe
datastructure produced a bug when I was just trying to translate existing logic suggests
to me that we actually are getting nontrivial safety from the compiler forcing
us to deal with cases explicitly, so I'm going to leave the current datatype
as-is for now.

Reviewed By: ndmitchell, grievejia

Differential Revision: D73144608

fbshipit-source-id: 3d2303dc36d66cb63eab4125840509db349c08f9