Simplify `TypeInfo` and `NarrowedAttrs`
Summary:
I noticed when implementing `join` (and to a lesser extent some other operations
like `add_narrow`, but especially `join`) that the datatypes in `TypeInfo` are
easier to line up with the actual logic if we make `attrs` an option, instead of
wrapping an option inside of it.

This allows us to make `join` return an option, where in the no-overlap case we get
`None`. It eliminates nesting in almost every helper of `NarrowedAttrs` (which is where
the most complexity lives), which makes things readable, and eliminates some very
strange looking matches in `NarrowedAttr::join` where we had to handle `None` and
`NarrowedAttrs(Some(None))` as equivalent.

Reviewed By: ndmitchell, grievejia

Differential Revision: D73144609

fbshipit-source-id: f12a93171e5568507a7e894a453e86ddfe8e60c6