Enable finding -stubs packages in site_package_path
Summary:
This diff enables finding `-stubs` packages when searching for a module. When `find_stub_packages = true`, we attempt to find a directory named `<package>-stubs` to start our search, and if not, then fall back to the regular package name.

A majority of this diff involves adding/updating the tests in `finder.rs`.

Reviewed By: ndmitchell, grievejia

Differential Revision: D73124782

fbshipit-source-id: ce8c9e2e6c9ffe9f46cc63c8b12ba8d9a3ee31e8