Rename function to `record_self_attr_assign`
Summary:
The old name was a bit misleading - if the term "bind" has any actual
meaning, it implies that we're binding a name that is in scope.

This operation does not bind a name in scope, it records an assignment
in a look-aside data structure where the class body handler can potentially
find it later, so naming it something other than "bind" makes things
clearer.

This change is partially motivated by the fact that we actually *do* need
to preform a real bind operation on attribute assignments to reflect
attribute narrowing, and having another think named "bind" in the mix makes
that more confusing than it needs to be.

Reviewed By: rchen152

Differential Revision: D73226176

fbshipit-source-id: 5c809fe6d85d2a45c1e5ea146882cb2a52b0e939