Add new test cases
Summary:
These are to track fixing the root cause of
https://github.com/facebook/pyrefly/issues/100
which is that we are not correctly accounting for the true
boolean structure when a narrow impacts multiple names, but instead just
dropping type information if we hit an `or` which causes a few issues.

The first test case is the one from the issue which is caused by negation; the
second tracks a case where with a more complex boolean, even in the positive
branch we will incorrectly drop type information on `x` due to the presence of
`y` in the same narrowing tree.

Reviewed By: ndmitchell

Differential Revision: D73235997

fbshipit-source-id: 3eceda3854397b9e295f3af2bbf158dbd0c45fd3