Make `And` start with a proper initial expr
Summary:
I am about to start using `Placeholder` narrow ops to correctly handle the
boolean algebra involved in multiple-name narrowing that can include
both `and` and `or` operators.

One issue I run into, though, is that our handling of `And` is currently taking
a shortcut where we start from an empty map instead of properly starting from
the leftmost expression, which leads to injecting nonsence Placeholders
if we don't fix this first.

This diff fixes the behavior, in the process making `and` and `or` handling
more similar. There is no associated behavior change.

Reviewed By: ndmitchell

Differential Revision: D73236003

fbshipit-source-id: fd0911acd6e87e5712c05b2f520838263d45f323