Rename truthy/falsy cases
Summary:
I am beginning work on https://github.com/facebook/pyrefly/issues/100,
whose root cause is that we cannot really simplify the boolean structure
of a narrow, because we need the ability to negate it as well as perform
other type algebra.

To solve the problem, I'm going to introduce an atomic narrow op that is
able to act as a placeholder where the value itself is not being narrowed,
but we are aware of the existence of a node in the boolean tree.

In the process of doing that though, the names `Truthy` and `Falsy` become
confusing, because as soon as we are tracking conditions that don't actually
narrow, these could potentially mean "a node in the boolean tree that is
known to be truthy / falsy" which is not what they are.

Renaming to `IsTruthy` / `IsFalsy` makes the names more obviously parallel
to the other narrow op names, and (at least in context) more clearly indicate
that we are saying "the value being narrowed is truthy".

Reviewed By: ndmitchell

Differential Revision: D73236001

fbshipit-source-id: 508f9c15cf2aaab5fa48599889714036a766870b