Deduplicate and vs or logic
Summary:
Use a single match and a closure to handle `and` vs `or`.

I'm not convinced the new code is much easier to *read* in the current form,
but I do think it is more bug-resistant: given that divergence between the two
cases made my narrowing fix in this stack more difficult to ship I think it's
probably worth deduplication.

Reviewed By: ndmitchell

Differential Revision: D73235998

fbshipit-source-id: cc5a55d44f466a7eac7393c0d56b05dc0d9b680e