Do not hack `from_single_narrow_op` using `And`
Summary:
Similar to the preceding diff: we are blocked from properly modeling `and` and `or`
operations against multiple names because parts of the narrowing code are relying
on an implementation detail of `And` (that you can start with an empty ops and just
tack stuff on) that isn't actually correct once we want to be able to negate
complex conditions.

So we need to move away from this: we should directly construct the correct value
in the `from_single_narrow_op` helper instead of trying to bootstrap it via
`And`.

Reviewed By: ndmitchell

Differential Revision: D73236002

fbshipit-source-id: 17c96281382eb553138ae31de38256cf004ab538