Do not hack multi-compare via `And`
Summary:
This diff fixes another case where coupled actual programming language logic
(how to narrow multi-comparison) to implementation details of `And` handling
that actually turn out to be incorrect when we have a complex multiple-name
narrowing expression, in ways that block fixing
https://github.com/facebook/pyrefly/issues/100

Instead of trying to start from empty NarrowOps and "and" our way to a result,
which makes the condition impossible to correctly negate once we are properly
tracking placeholders, we ought to be directly computing the first condition
and then only "and"-ing them if there are more conditions.

The resulting code is simpler but less efficient, since it is no longer
convenient to pre-compute the subjects (we could do it, but it would require
a whole slew of additional helpers). Multi-compares are probably pretty rare
in practice, and extracting subjects is dirt cheap, so I really don't think
this matters and I think we should live with it unless it comes up in a
profiling run.

Reviewed By: ndmitchell

Differential Revision: D73235995

fbshipit-source-id: e339e4b4e63e4704fd53210c25c1338fa6f47e43