Don't commit the transaction in one-shot mode
Summary:
Time to check 100K modules for the transaction/loads creation bit drops from 191ms to 26ms.

Requires moving to set the subscriber, since we want to kill the progress bar once we finish, but before we print out the data. And we no longer control the lifetime of the transaction.

Reviewed By: SamChou19815

Differential Revision: D73310227

fbshipit-source-id: 021989cf1e73aa57480b7b19b08f5c36d6f0c940