Add assignment tests
Summary:
I was originally thinking we needed narrows to propagate, but then
I realized that is not even sound - due to mutation, and the restricted
scope of type-level analysis (we're not going to track identities of
all narrows in a scope and wipe them all on a mutation), that isn't
a good idea and Pyre1 doesn't do it.

So, adding a test that demonstrates what does work (the type of an expression
assigned to a name will reflect all live narrows) and what does not (we will not
preserve any trees of narrows *underneath* a name), with comments explaining
why this is expected.

Reviewed By: ndmitchell

Differential Revision: D73371128

fbshipit-source-id: 0a7ec42fee70bbca7bf68f26dd570718fef6b009