Move `Binding::Expr` back to `binding_to_type`
Summary:
I originally thought that I would need this to support `TypeInfo`, but this is
not the case - the thought that went into the previous test shows that we only
want to propagate the (possibly narrowed) outermost type of an expression
into assignments and downstream type-level operations, not the tree of attribute
narrows.

That means this belongs in `binding_to_type` rather than `binding_to_type_info`.

Reviewed By: ndmitchell

Differential Revision: D73371127

fbshipit-source-id: d82f1a103dbe7568d269315476e4839bbfe6d4b2