Fix callable parameter subtyping when RHS param list is empty
Summary:
When performing subtyping query between `Callable[ARG_LIST, X]` and `Callable[[], X]`, we check if `ARG_LIST` starts with `*args` or `**kwrags` -- if that's the case, Pyrefly immediately conclude that LHS is a subtype of RHS.

This kind of early-return-true logic works for `**kwargs` but not for `*args`. This is because there might still be more elements in `ARG_LIST` after the `*args` that need to be consumed, and if those args don't have a default value, LHS won't be compatible with an empty arg list (see added test `varargs_bad`). `**kwargs` doesn't have the same problem because Python syntax guarantees that no more params can be declared after `**kwargs`.

This diff makes it such that we we see `*args` (or `**kwargs` or any other kind of optional param) in `ARG_LIST`, we move ahead and inspect the next element in the arg list as opposed to immediately return true. This would guarantee that when RHS param list is empty, we always inspect everything in LHS `ARG_LIST` to make sure that none of the remaining elements is skippable.

Reviewed By: rchen152

Differential Revision: D73226130

fbshipit-source-id: a4146ef3e2657d43e5d440d7e1587c85e3ff1adc