keep opacity the same when button is disabled
Summary: The button disabled style currently has opacity set to 0.5, A button is currently only disabled when it is in the running state. Given that running has it's own visual indicator already (green color), I'm going to remove the opacity change since this seems redundant. The hover style is redundant as well since this should already exist for both the default button and running button states.

Reviewed By: MaggieMoss

Differential Revision: D73374439

fbshipit-source-id: ab206740b4f2a2061446b19b011551d7f64143b4