Rename isRunning to isClicked in MonacoEditorButton
Summary: rename variable to be more accurate. `isSuccessful` is not necessarily true since we set this to true when the button is clicked, and false whenever the action completes, regardless of if it succeeds or not. `isRunning` is a much better name here. Update other instances of "success" to "running" as well, along with adding a comment for share url button indicating why we reuse the `runningLabel` as a "success" label

Reviewed By: MaggieMoss

Differential Revision: D73373851

fbshipit-source-id: 1ddeee1c1485f8ae9378b56f9e47bb5bc2d20c05