Fix iteration over class objects
Summary:
When iterating over a class object, we need to check its metaclass for `__iter__` and `__getitem__`, not the class itself.

I was hoping this would (finally!) fix iteration over Enum classes, but that's still broken because `EnumMeta.__iter__` is generic =( I updated the bug=... comment on the corresponding test.

Reviewed By: ndmitchell

Differential Revision: D73293425

fbshipit-source-id: 6bc4b29dc107be40f2aedc33bf05de3071598036