Pass around an Instance rather than a ClassType for instance method binding
Summary:
With this, we correctly bind TypedDict methods to Type::TypedDict rather than Type::ClassType. This also allows us to fix a bit of weirdness where we were forced to instantiate "self" param types incorrectly to match the previously incorrect method binding.

This isn't a pure refactor because there's one place (as_instance_attribute, line 866 in class_field.rs) where we now refuse to create a DescriptorBase inside a TypedDict. I'd be very surprised if anyone's trying to do this in the first place.

Context: I'm trying to fix some TypedDict bugs. This has turned out to be more challenging than expected, due to the fact that the attribute handling code only operates on ClassTypes. I'm first doing some refactoring to change that.

Reviewed By: ndmitchell

Differential Revision: D73281424

fbshipit-source-id: c817e2fc73b0ee299e62aef6bf82cbf07bb22d80