Use Instance instead of ClassType in a bunch more places in class_field.rs
Summary:
Bubbles use of Instance all the way up to the public get_dunder_init method. This will make it easy for me to add a flavor of get_dunder_init for TypedDict.

Context: I'm trying to fix some TypedDict bugs. This has turned out to be more challenging than expected, due to the fact that the attribute handling code only operates on ClassTypes. I'm first doing some refactoring to change that.

Reviewed By: ndmitchell

Differential Revision: D73281380

fbshipit-source-id: 38581ee6249be99f857bd0b95d8f1f67ef818849