Move ClassType::instantiate_member into class_field.rs and make it able to take any specialized class
Summary:
The goal is (again) to make it so that this function can operate on TypedDict.

Context: I'm trying to fix some TypedDict bugs. This has turned out to be more challenging than expected, due to the fact that the attribute handling code only operates on ClassTypes. I'm first doing some refactoring to change that.

Reviewed By: ndmitchell

Differential Revision: D73276921

fbshipit-source-id: 1521d034c8df800212bfe3ba85544e8d693f407b