Add tests for error suppressions
Summary:
Round 1 of basic tests to make sure we're not deleting people's code when we remove unused suppressions.

I'm intentionally leaving empty lines when we remove suppressions to ensure anything else in the scripts that requires a line number is still reported correctly and because we currently rely on a formatter to clean these up after we remove suppressions in pyre1.

Reviewed By: ndmitchell

Differential Revision: D73213174

fbshipit-source-id: 781e0dc84656d2f30d6ec5a646c92bab8137cae7