Add tests validating behavior of glob matching (#126)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/126

This diff makes no functional changes. It only fixes a few lines in documentation (which I plan to expand on during the docathon) and adds tests for existing functionality, some of which has bugs.

Over the next few diffs, I'll uncomment all of the tests below as I get things working, but the tests below show the state of how I think globs should be working.

We also don't have a way to verify that filesystem matching is working as expected, so these tests cover that as well.

Reviewed By: grievejia

Differential Revision: D73378283

fbshipit-source-id: 914dcf3fcb545bb61d697f0fdd5cba0146905a09