Don't produce duplicate bad-annotation-target errors
Summary:
Much as in the previous diff, we don't need to produce a duplciate error
on bad annotated assignment targets that Ruff already catches, which it
does with error messages that are as good or better than the ones we
have coming directly from Pyrefly.

The one case we do want to catch is Subscript, and the error message
can be improved in that case.

There is one case where we are actually getting a duplicate error still,
which is the Starred case; I think that's probably okay, it's an edge case
and my main complaint was that we were doing *extra* work and adding complexity
to our code just to produce duplicate errors, here it would be a little
challenging to dedup. I left a TODO to think about it.

Reviewed By: grievejia

Differential Revision: D73398827

fbshipit-source-id: 78892f5b3b246f3fe87a89be00b9e2b2feb2b2f9