Ensure exprs inside of `bind_target`
Summary:
Up until now, Pyrefly has been resovling `bind_target`, which will bind
all the names assigned in a target, and *then* call `ensure_expr` on the target.

This sort of works because we never actually use an assigned name, but conceptually
it is a bug: we're time-traveling so that the target resolution potentially depends
on the result of the resolution.

Once attribute narrowing is supported, this bug breaks Pyrefly because now
every attribute assignment tries to produce the resulting `TypeInfo`, but then
our time-traveling binder will say the type *before* narrowing is the result of
the narrow, which leads to an unresolvable Var and winds up turning the type
into Any (this is pretty severe because it would happen on *every attribute
assign!*, so it's pretty important to fix beforehand).

The solution is to ensure_expr only where we actually need a bound expression
- Name targets generally *don't* need a bound value, but there's a special
  case for `AugAssign` (which only applies to top-level names, we don't have to
	deal with it in `bind_unpacking) where they do.
- Subscript and attribute targets always need the base to be bound, so we always
  ensure the base.
- Invalid assignment targets don't *really* matter (they are a type error coming
  from the parser anyway), but we do call `ensure_expr` anyway because we want
	to be able to catch the use of undefined names, e.g. if `x` is undefined
	in `x + x = 42`.

Reviewed By: ndmitchell

Differential Revision: D73407433

fbshipit-source-id: 0035ebac69d852a340af172ebeccc3e0085f284d