Implement site package path errors with py.typed behavior non-`*-stubs`
Summary:
# This diff...

Updates `find_module_in_search_path()` to add the logic described below for searching non-`*-stubs` modules.

This involves iterating through all top-level `site_package_path` entries and filtering them to only those that match the first component of the path (a `<library>` directory under a site package path).

If we found a `partial\n` `py.typed` in `site_package_path` entries with `-stubs`, then search all packages we found in the step above. Otherwise, only search packages that have a `py.typed` file.

We then check each of those to see if they have the import we're looking for. If they do, then we're good, return that and be done.

If not, then we failed to find the package, so return `None`.

# This Stack...

Over the next couple of diffs, I'll be adding support for erroring on missing `*-stubs` or `py.typed` files. The [type checker library specification](https://typing.python.org/en/latest/spec/distributing.html#packaging-type-information) has rules around when third party packages can be used in type checking, so we should add these rules when trying to import something from `site_package_path`.

Order of diffs:
1. Add a new `use_untyped_imports` flag, which can be used to override `py.typed` behavior.
2. Add a new `FindError` type that gives a helpful error message about ignoring `py.typed` behavior with `use_untyped_imports`.
3. Add functionality to get and read a `py.typed` file for a `FindResult`.
4. Split `find_module_in_search_path` into `continue_find_module` to reuse the logic we need to figure out if there's a first component match instead of duplicating the logic.
5. Add tests that we expect to update in the following two diffs.
6. Update `find_module_in_search_path` to perform the logic in bullet (1) below.
7. Update `find_module_in_search_path` to perform the logic in bullet (2) below. (this diff)
8. Update `find_module_in_search_path` to return a `FindResult` if it's possible we didn't check usable modules because of a missing `py.typed` with `partial\n`.
9. Use `use_untyped_imports` in `find_module_in_search_path`

I've understood the specification to be this, and my following implementation will try to emulate it:
1. search for any -stubs libraries. If they exist, then only use them and stop the search, unless any of them contain a py.typed with `partial\n`, in which case, continue with the following steps. We should not require a `py.typed` file for these imports, which is specified in the docs
`For stub-only packages adding a py.typed marker is not needed since the name *-stubs is enough to indicate it is a source of typing information.`
2. search for any libraries that match the first component in the name we're searching for. filter out any that don't have a `py.typed` (unless `partial\n` was found in `*-stubs`) and use those to search for imports, and resolve imports normally. If none were found, but there were some that don't have a `py.typed`, then return an error that signifies packages were found, but we can't import them because there's a missing `py.typed` file, but there's a flag you can use to override that.

A python-like approach might look something like this:
```
has_partial = false
# matching_stub_libraries are filtered site_package_path entries
# where the first component of the entry matches the first component
# of the import + `-stubs`
for stub_library in matching_stub_libraries:
  import = stub_library.find_matching_import()
  if import:
    return Some(import)
  has_partial |= stub_library.has_partial_py_typed()

if len(matching_stub_libraries) > 0 and not has_partial:
  return None

# mathing_libraries are the same as above, but without `-stubs` (so
# matching the first component of the import exactly)
for library in matching_libraries:
  if not (use_untyped_imports or has_partial or library.has_py_typed()):
    continue
  import = library.find_matching_import()
  if import:
    return Some(import)

if len(matching_libraries) > 0 and not has_partial:
  return FindError::NoPyTyped

return None
```

Reviewed By: grievejia

Differential Revision: D73284354

fbshipit-source-id: 16794efc0d96b57f5bea01308e5ff9f7d78f4ae0