Populate FAQ in sidebar + publish to public
Summary: I created a placeholder for the faq originally in D72993480 in preparation for the docathon. Aaron landed the actual FAQ in D73111469 but this isn't linked to the sidebar. This diff removes the empty FAQ and links the actual FAQ in the sidebar. This also publishes this to public. I chatted with Aaron offline and we are good to release this publicly.

Reviewed By: MaggieMoss

Differential Revision: D73449022

fbshipit-source-id: f0007c3522a8fe0fd449cfc4fda21129fc816a17