Return FindError::NoSource when module source can't be found for site package path
Summary:
Pyright has a `reportMissingModuleSource` config, which returns an error when there's a `*-stubs` library found with no backing source library. It seems like a great idea, so I'll be implementing it here too.

This diff implements the functionality to error when a found `*-stubs` module has no missing source, using the newly created `ignore_missing_source` config option to not surface the error when requested.

The idea around this is similar to the previous approach for finding a site package, with the following changes:
1. find a `-stubs` library import, if it exists
2. **new**: if `ignore_missing_source` is true, return immediately
3. find non-`stubs` library imports
4. If we found no partial py.typed file, but we found `*-stubs` packages, then return `None`
5: **new**: if there are no non-`stubs libraries, but we found a `-stubs` import, then return a `FindError::NoSource`
6. **new**: if there are non-stubs libraries found, return the previously found stubs result (if present), otherwise, return the non-stubs library
7. Return some kind of `NotFound` error, depending on previously implemented logic

## Open question

There is a remaining question about what we should count as a valid non-stubs library when trying to figure out if a source exists. Do we:
1. count any non-stubs library found as being a source, or do we
2. require the non-stubs library to have the exact import we're looking for

I've gone with #1 right now, since that seems most lenient, and probably good enough for 99% of cases.

Reviewed By: grievejia

Differential Revision: D73412299

fbshipit-source-id: 4667664f5a0bb839f73ec3216694fd4fb4cfd001