Add module name to FindError::NotFound
Summary:
Pyright has a `reportMissingModuleSource` config, which returns an error when there's a `*-stubs` library found with no backing source library. It seems like a great idea, so I'll be implementing it here too.

This diff chagnes the `FindError::NotFound` variant to also take in the module it's acting on, since in every situation it's used, we have to pass that information along. This makes the `FindError::display()` function easier to use, since each variant knows what it should be returning as its description.

Reviewed By: grievejia

Differential Revision: D73457138

fbshipit-source-id: 7ad652e058f931f251b3b32dbb80644f43446efe