Preserve directory structure when trimming typeshed
Summary:
Keep typeshed's stdlib stubs inside typeshed/stdlib/ instead of moving them up to typeshed/. To do this, I moved the typeshed files themselves, updated the fetch_typeshed script to stop removing the stdlib/ path component, and tweaked bundled.rs to remove the stdlib/ component before obtaining module names.

Context: migeed-z and I are looking into getting pyrefly into typeshed CI. As a first step, we want to be able to use a custom typeshed rather than pyrefly's bundled one. This is easier if the directory structure is the same.

Looking at D67432083, the purpose of trimming typeshed when bundling it is to reduce the size; this change shouldn't significantly affect the size.

Note that I `sl mv`ed the files to preserve their history, but afterwards, I ran fetch_typeshed to make sure its output exactly matched the result of the move.

Reviewed By: grievejia

Differential Revision: D73475176

fbshipit-source-id: 1cad9587cd09acf101d84d853ba640ebc0cc0d97