Recognize all rust-supported globs and escape root
Summary:
Since we can match on `*`, `**`, and `?`, why not expand it to all glob patterns? This diff lets us find the start of a character range by searching for `[`.

We also escape the root so that weird filesystem stuff won't interfere with the root here (I'm thinking buck might do some interesting stuff), but we can still do patterns within the project.

Reviewed By: Solumin

Differential Revision: D73413268

fbshipit-source-id: 8b1a728cfedd7f9b614f30da1dbb8cb557677db5