Handle matching immediate directory entries when glob ends with '**'
Summary:
When listing files from a glob pattern, we run into a tricky bug with the glob library we're using when a file ends in a `**` wildcard. In that case, we only list the directories found, and then iterate through them manually with `resolve_dir()`. This means any files under the directory will not be found (e.g. with a pattern `mydir/**`, we'll list any directories under `mydir`, and files under those directories, but not files immediately within `mydir`).

We can solve this issue by adding a `*` wildcard to any patterns we load.

Reviewed By: grievejia

Differential Revision: D73400444

fbshipit-source-id: cda19dc06a1269fd63622d46aa32c2fd94f08d09