Avoid artifically altering lambda param lists when contextually typing lambda expressions
Summary:
D69830147 added contextual typing to lambda parameters. The way it works is that when seeing a lambda expression like `lambda x, y: ...` in a context that expects type `U`, we create a callable type `C` with two vars for parameter types and a var for return type, and does subtyping check `C <: U` to force all the vars in `C`. Then we use `C` as a hint to type check the lambda expression.

This is all nice and good. But there's one trick introduced in that diff (which already [got pointed out during code review](https://www.internalfb.com/diff/D69830147?dst_version_fbid=2368630996815546&transaction_fbid=1719400868617120)) I think would be debatable: when creating the callable type `C`, we don't faithfully respect the number of parameters in the lambda expressions. So for `lambda x, y: ...`, instead of having `C` to be the type `(var_x, var_y) -> var_ret`, we alter the param list a little bit to make it more lenient and can take arbitrary amount of arguments: `(var_x = ..., var_y = ..., *args: var_a) -> var_ret`.

The justification for doing that, according to the explanation, is to "give more stability for transiently-invalid code", which IIUC is to address a hypothetical UX concern. Unfortunately, the trick actually affect type checker correctness, as demonstrated by [this github issue](https://github.com/facebook/pyrefly/issues/117):
```
def foo[T](x: T) -> T: ...
foo(lambda: None)
```
Here, the expected type is `T` for the lambda expression. What's going to happen here is that we first create the contextual callable type `C = (var_x=..., *args: var_a) -> var_ret`, and does `is_subset_eq(C, T)`. Note that this time, `T` is a typevar -- so all the `is_subset_eq` does is to enforce that `T = C`. But this creates a problem because the program no longer type checks: Now that `T = C`, plus the type of the lambda expression must now be a subtype of `T`, we now start to enforce that the type of the lambda expression must be a subtype of `C`. Yet if we inspect `C` -- it requires that the lambda must be lenient enough to be able to take *arbitrary* amount of arguments. That's certainly impossible since our lambda expression actually can only take *no* argument at all!

The crux of the problem seems to be that when creating contextual types against generic types, it's incorrect to artificially alter those created contextual types in a way that's not equivalent to the original type of the expression being checked, because those generic types may be "reflected back" (via, e.g. is_subset_eq query side effects) to somewhere else in the program where the equivalence is expected. Given that having the original behavior is just about the UX whereas not having the behavior is actually required for correctness, I would propose that we go with not having the behavior now, and if we really want to address the UX issue we can definitely come back to it later but it needs to be done in a way that does not affect correctness.

Fixes github issue #117.

Reviewed By: ndmitchell

Differential Revision: D73456439

fbshipit-source-id: 80d5bedfe804151ba1533352ad2493ce68da8cfe